Add a progress bar which tracks ratio of sites supporting BCH
