+++ .gitignore
@@ -11,4 +11,3 @@ Gemfile.lock
 *.suo
 *.pyproj
 scripts/python/output/
-/_includes/count_support.html
