+++ _includes/count_support.html
@@ -1,0 +1,1 @@
+ <div class="ui container progress bch-progress" data-value="107" data-total="962" id="bch-count">   <div class="bar">     <div class="progress"></div>   </div>   <div class="label" data-tooltip="Please help to improve this ratio by sending tweets to, or writing on the wall of each website listed above that does not yet support Bitcoin Cash." data-position="bottom center" data-inverted=""><b>107</b> out of <b>962</b> websites listed support Bitcoin Cash.</div> </div>
